requirements.txt
pandas
numpy
joblib
scikit-learn
websocket-client
ta
requests
